requirements.txt
Adafruit-Blinka==8.49.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-dht==4.0.5
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.1
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
anyio==4.6.2.post1
binho-host-adapter==0.1.6
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
mariadb==1.1.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.5
packaging==24.1
pika==1.3.2
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyftdi==0.55.4
Pygments==2.18.0
PyMySQL==1.1.1
pyserial==3.5
python-dotenv==1.0.1
python-multipart==0.0.16
pyusb==1.2.1
PyYAML==6.0.2
rich==13.9.3
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.2
sysv-ipc==1.1.0
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
